requirements.txt
tensorflow==2.13.0
python-dotenv
pandas